pricedb API
./BitcoinDarkd SuperNET
'{"requestType":"pricedb","exchange":"bter","base":"BTCD","rel":"BTC"}'

turned out to be a lot of little details and I still have some DB
deadlock cases to fix, but if you have only one active pair it seems to
be working. not sure if all the data is getting into the DB, but the
dbput is coming back without error.

so what is this pricedb API? it creates an entry in a small db file
which is just a list of exchanges and trading pair that is active. The
cool part is that for all active pairs, it automatically polls the
prices from the exchange and creates a database file locally. This way,
over time you will get price data to analyze, but more near term a way
to see the current market price.

The exchanges that are supported are: "NXT", "bter", "bittrex",
"cryptsy", "poloniex", "mintpal"
for NXT, only the currencies with MGW assets are supported, for the
others, any trading pair the exchange supports is supported. [I think
cryptsy only works with a hardcoded list for now]